+++ src/collection/requirements.txt
@@ -1,0 +1,23 @@
+attrs==25.4.0
+beautifulsoup4==4.14.3
+certifi==2026.2.25
+charset-normalizer==3.4.5
+exceptiongroup==1.3.1
+h11==0.16.0
+idna==3.11
+lxml==6.0.2
+outcome==1.3.0.post0
+PySocks==1.7.1
+requests==2.32.5
+selenium==4.36.0
+sniffio==1.3.1
+sortedcontainers==2.4.0
+soupsieve==2.8.3
+trio==0.31.0
+trio-websocket==0.12.2
+typing_extensions==4.15.0
+undetected-chromedriver==3.5.5
+urllib3==2.6.3
+websocket-client==1.9.0
+websockets==15.0.1
+wsproto==1.2.0
