--- src/salty_polaris/__init__.py
@@ -1,5 +1,0 @@
-"""salty-polaris: VesselFinder port scraper."""
-
-from .scraper import VesselFinderScraper, VesselRecord
-
-__all__ = ["VesselFinderScraper", "VesselRecord"]
